Rework attack magics
Description:
- Use minimal space for attack tables.
- Enable multi-threading through global shared
attack tables.

+++ inc/types.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-using move_t = std::uint32_t;
+using move_t = std::uint16_t;
 using bitboard_t = std::uint64_t;
 using square_t = std::uint8_t;
 using score_t = std::int16_t;
